requirements.txt
TA-Lib-binary==0.4.21
fastapi==0.75.2
uvicorn==0.17.6
pandas==1.4.2
